Provide role scopes as csv to align with config
Config already provides its scopes as csv.

Use-case:

I want to set `allowed_parameters` in the policy for `scopes`. However, this is only string based. When using this with terraform and supplying the scopes string based, there will be a constant state drift as the response is different (an array vs a string).